+++ src/main/java/com/vuthy/mobilebankingapi/dto/AccountResponse.java
@@ -12,6 +12,6 @@ public record AccountResponse(
         String actCurrency,
         BigDecimal balance,
         Boolean isDeleted,
-        Double overLimit
+        BigDecimal overLimit
 ) {
 }
