+++ src/main/java/com/vuthy/mobilebankingapi/service/Impl/SegmentServiceImpl.java
@@ -8,6 +8,7 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 
@@ -21,16 +22,16 @@ public class SegmentServiceImpl implements SegmentService {
     @Transactional
     public void initDefaultSegments() {
         List<Segment> defaultSegments = Arrays.asList(
-                createSegment("Regular", 5000),
-                createSegment("Silver", 10000),
-                createSegment("Gold", 50000)
+                createSegment("Regular", BigDecimal.valueOf(5000)),
+                createSegment("Silver", BigDecimal.valueOf(10000)),
+                createSegment("Gold", BigDecimal.valueOf(50000))
         );
 
         segmentRepository.saveAll(defaultSegments);
     }
 
     @Override
-    public Segment createSegment(String name, Integer overLimit) {
+    public Segment createSegment(String name, BigDecimal overLimit) {
         Segment segment = new Segment();
         segment.setName(name);
         segment.setOverLimit(overLimit);
